requirements.txt
claude-agent-sdk
pydantic
python-dotenv
